# More efficient test for binary files in matrix.comp.