# Pull Request Summary
Provides a more efficient test for binary files when running matrix.comp

## Description
Currently, matrix.comp recursively checks all files in a work directory using the command
```
 binaryfiles=`grep . -r * | grep 'Binary file' | sed -e "s/^Binary file //" -e "s/ matches$//"` 
```

This is very slow as:
1. It will unnecessarily read the entirety of any text files (`.` matches "any character")
2. It recurses into the CMake `build*` and `exe` directories which have a **lot** of files in.

A better solution is to use the standard linux [`file`](https://www.man7.org/linux/man-pages/man1/file.1.html) command to check whether individual files are "plain text" files or binary. This is not only a much quicker and more efficient test, but it will only be called when needed (as apposed on a complete recursive listing of the directory)

I assuming a fairly modern implementation of the `file` command that accepts the `-i` or `--mime` flags for robustness purposes. The test is encapsulated as a function at the top of the `matrix.comp` script to facilitate modification if needed.

#### Performance
Some results of running new vs old version of matrix.comp:
 - `ww3_tp2.2` (10 work directories):
   - original version: 3m 7s
   - new version: **9s**
 - `ts1` (11 work directories):
   - original version 2m2s
   - new version:  **7s** 
  - `tp2.9` (20 work directories):
    - 6m 9s
    - **20s**

_Note: When testing performance, it is important to **not** rerun `matix.comp` twice on the **same** regression test directory as the file system will likely have cached the files and the timings will be much quicker the second time._

I make performance comparisons by testing in two separate WW3 regtest trees (and testing the regtests against themselves)

### Issue(s) addressed
Fixes #1031 

### Commit Message
More efficient test for binary files in matrix.comp

### Check list  

<!-- After creating the PR you can check each of the items below that have been completed -->

- [x] Branch is up to date with the authoritative repository (NOAA-EMC) develop branch. 
- [x] Checked the [checklist for a developer submitting to develop](https://github.com/NOAA-EMC/WW3/wiki/Code-Management#checklist-for-a-developer-submitting-to-develop). 
- [ ] If a version number update is required, checked the [updating version number](https://github.com/NOAA-EMC/WW3/wiki/Code-Management#checklist-for-updating-version-number) checklist. 
- [ ] If a new feature was added, a regression test for testing the new feature is added. 
 
### Testing

* How were these changes tested? **Comparison of regression test output; diffs out matrixCompFull.txt are same**
* Are the changes covered by regression tests? (If not, why? Do new tests need to be added?) **N/A**
* Have the matrix regression tests been run (if yes, please note HPC and compiler)? **N/A**
* Please indicate the expected changes in the regression test output, (Note the [list](https://github.com/NOAA-EMC/WW3/wiki/How-to-use-matrix.comp-to-compare-regtests-with-develop#4-look-at-results) of known non-identical tests.) **N/A**