@JessicaMeixner-NOAA 2023-07-07
> As part of this PR, I wonder if others would also agree to the removal of the "do not compare build or exe directory" messages to STDOUT?
> 
> ```
>  testing case: ww3_tp2.9; run: ./work_PR2_UQ_a_MPI
> do not compare build or exe directories build
> do not compare build or exe directories build_SHRD
> do not compare build or exe directories exe
> do not compare build or exe directories build
> do not compare build or exe directories build_SHRD
> do not compare build or exe directories exe
> ```
> 
> these get printed multiple times for each test and dominate the output to the screen.

@ukmo-ccbunney I'm okay with that