@MatthewMasarik-NOAA 2023-07-07
I'm also OK with removing the repeated prints of "do not compare build or exe directory".