@ukmo-ccbunney 2023-07-07
Timing results from running `matrix.comp` on a full regression test matrix:
  - old code: 272m (~4.5 hours)
  - new code 63m

So - not as fast a speedup as the results for individual regtests show above (not sure why?), but still a significant improvement!