@ukmo-ccbunney 2023-07-07
> I'm also OK with removing the repeated prints of "do not compare build or exe directory".

I've pushed a commit to remove this message.