@JessicaMeixner-NOAA 2023-07-07
Thanks @ukmo-ccbunney !   I am running tests to do comparisons between the old and the new on our machines. Those should be finished soon and then I can submit the two matrix.comps for comparisons.   I'll let you know how it goes!