@ukmo-ccbunney 2023-07-11
> This improved the speed from 80min to 53 min on hera for me. Thank you for this speed up @ukmo-ccbunney !

I'm impressed they were running in 80 minutes before! I think the regtests really stress the filesystem on our HPC... it takes over 4.5 hours for me to run them!