@ukmo-ccbunney 2023-07-12
> @ukmo-ccbunney well, i just had a set of matrix.comp take 4.5 hours.... with the updates, so .... I wonder if we need to rethink what we're comparing and how we're comparing it to relieve some of the stress. 4.5 hours is crazy, but I get that once in a blue moon these days. I usually have the tests get stuck on the ufs tests at the end if something does take a long time.

Wow. 4.5 hours with this update? I wonder if that is just some spurious filesystem contention that you are experiencing.
This update has certainly helped speed thiings up a lot on my HPC, but every system is different! 
I think it is definitely high time we looked at the matrix.comp script and what/how it is comparing.